pyproject.toml
[tool.poetry]
name = "data-parsing"
version = "0.1.0"
description = ""
authors = ["Egor <egorkononec64@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
supabase = "^2.5.1"
botasaurus = "^4.0.47"
peewee = "^3.17.6"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.5"
npm = "^0.1.1"
twocaptcha = "^0.0.1"
pinecone-client = "^5.0.0"
loguru = "^0.7.2"


[tool.poetry.scripts]
dev = "main:start_dev"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
